removed the JSON dependency
