+++ api/user-data.json
@@ -1,1 +1,1 @@
-{"absentee_check": "", "birth_day": "", "birth_month": "", "birth_year": "", "conservative_check": "", "democratic_check": "", "dmv": "", "election_worker_check": "", "email": "", "female": "", "first_name": "", "green_party_check": "", "id_dmv_check": "", "id_ssn_check": "", "independence_check": "", "last_name": "", "mail_address": "", "mail_city": "", "mail_po": "", "mail_zip": "", "male": "", "middle_initial": "", "no_id_check": "", "no_party_check": "", "o_address": "", "o_apt": "", "o_birth_day": "", "o_birth_month": "", "o_birth_year": "", "o_city": "", "o_eye_color": "", "o_female": "", "o_first_name": "", "o_height_feet": "", "o_height_inches": "", "o_last_name": "", "o_male": "", "o_suffix": "", "o_zip": "", "older_than_18_no": "", "older_than_18_yes": "", "other_party": "", "other_party_check": "", "reform_check": "", "republican_check": "", "resi_address": "", "resi_apt#": "", "resi_city": "", "resi_county": "", "resi_zip": "", "return_address_apt": "", "return_city_state_zip": "", "return_name": "", "send_address": "", "send_address2": "", "send_city_state_zip": "", "send_county": "", "ssn": "", "suffix": "", "tel_area_code": "", "tel_first3": "", "tel_last4": "", "us_citizen_no": "", "us_citizen_yes": "X", "vh_address": "", "vh_county": "", "vh_name": "", "vh_year": "", "voted_before_no": "", "voted_before_yes": "", "women_equality_check": "", "working_family_check": ""}
+{ "absentee_check": "", "birth_day": "", "birth_month": "", "birth_year": "", "conservative_check": "", "democratic_check": "", "dmv": "", "election_worker_check": "", "email": "", "female": "", "first_name": "", "green_party_check": "", "id_dmv_check": "", "id_ssn_check": "", "independence_check": "", "last_name": "", "mail_address": "", "mail_city": "", "mail_po": "", "mail_zip": "", "male": "", "middle_initial": "", "no_id_check": "", "no_party_check": "", "o_address": "", "o_apt": "", "o_birth_day": "", "o_birth_month": "", "o_birth_year": "", "o_city": "", "o_eye_color": "", "o_female": "", "o_first_name": "", "o_height_feet": "", "o_height_inches": "", "o_last_name": "", "o_male": "", "o_suffix": "", "o_zip": "", "older_than_18_no": "", "older_than_18_yes": "", "other_party": "", "other_party_check": "", "reform_check": "", "republican_check": "", "resi_address": "", "resi_apt#": "", "resi_city": "", "resi_county": "", "resi_zip": "", "return_address_apt": "", "return_city_state_zip": "", "return_name": "", "send_address": "", "send_address2": "", "send_city_state_zip": "", "send_county": "", "ssn": "", "suffix": "", "tel_area_code": "", "tel_first3": "", "tel_last4": "", "us_citizen_no": "", "us_citizen_yes": "", "vh_address": "", "vh_county": "", "vh_name": "", "vh_year": "", "voted_before_no": "", "voted_before_yes": "", "women_equality_check": "", "working_family_check": "" }
